+++ core/src/main/java/io/substrait/function/ParameterizedTypeVisitor.java
@@ -15,6 +15,8 @@ public interface ParameterizedTypeVisitor<R, E extends Throwable> extends TypeVi
 
   R visit(ParameterizedType.IntervalCompound expr) throws E;
 
+  R visit(ParameterizedType.PrecisionTime expr) throws E;
+
   R visit(ParameterizedType.PrecisionTimestamp expr) throws E;
 
   R visit(ParameterizedType.PrecisionTimestampTZ expr) throws E;
@@ -51,6 +53,11 @@ public interface ParameterizedTypeVisitor<R, E extends Throwable> extends TypeVi
 
     @Override
     public R visit(ParameterizedType.Decimal expr) throws E {
+      throw t();
+    }
+
+    @Override
+    public R visit(ParameterizedType.PrecisionTime expr) throws E {
       throw t();
     }